refactor(client/review) Component Review Added
