+++ client/src/scss/components/_productDetails.scss
@@ -129,76 +129,3 @@
     font-size: 20px;
     font-weight: bold;
 }
-
-.review-cont {
-    display: flex; 
-    margin-left: 20%;   
-    margin-top: 50px;
-    white-space: pre;
-    width: 400px;
-}
-
-.reviewTitle {
-    font-size: 30px;
-    font-weight: bold;
-    padding-bottom: 10px;
-}
-
-.reviewComment {
-    display: flex;
-    font-size: 18px;
-    padding-bottom: 10px;
-}
-
-.add-review {
-    font-size: 30px;
-    font-weight: bold;
-    padding-left: 100px;
-    margin: 0;
-}
-
-.nameAndComment {
-    padding-left: 10px;
-}
-
-.reviewForm {
-    margin: 0;
-}
-
-.noReviewsYet {
-    font-size: 18px;
-    white-space: pre;
-    margin: 0;
-}
-
-.inputReviewComment {
-    width: 300px;
-    height: 100px;
-    margin:0;
-}
-
-.divReviewButton {
-    align-content: end;
-    margin-left: 16px;
-    width: 100px;
-}
-
-.spanAddReview {
-    font-size: 18px;
-    font-weight: normal;
-    padding-right: 10px;
-    align-content: flex-start;
-}
-
-.divAddReview {
-    display: flex;
-    align-content: flex-start;
-}
-
-.inputSubmit {
-    align-items:flex-start;
-    margin:10px 0;
-}
-
-
-
